APIs/math.tools/1.5/python-aiohttp-client/openapi_server/controllers/number_checks_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server import util


async def numbers_is_cube_get(request: web.Request, number=None) -> web.Response:
    """numbers_is_cube_get

    Checks whether a given number is a cube number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_is_palindrome_get(request: web.Request, number=None) -> web.Response:
    """numbers_is_palindrome_get

    Checks whether a given number is a palindrome number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_is_square_get(request: web.Request, number=None) -> web.Response:
    """numbers_is_square_get

    Checks whether a given number is a square number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_is_triangle_get(request: web.Request, number=None) -> web.Response:
    """numbers_is_triangle_get

    Checks whether a given number is a triangle number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_prime_is_fermat_prime_get_0(request: web.Request, number=None) -> web.Response:
    """numbers_prime_is_fermat_prime_get_0

    Checks whether a given number is a known fermat prime number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_prime_is_fibonacci_prime_get_0(request: web.Request, number=None) -> web.Response:
    """numbers_prime_is_fibonacci_prime_get_0

    Checks whether a given number is a known fibonacci prime number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_prime_is_mersenne_prime_get_0(request: web.Request, number=None) -> web.Response:
    """numbers_prime_is_mersenne_prime_get_0

    Checks whether a given number is a known mersenne prime number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_prime_is_partition_prime_get_0(request: web.Request, number=None) -> web.Response:
    """numbers_prime_is_partition_prime_get_0

    Checks whether a given number is a known partition prime number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_prime_is_pell_prime_get_0(request: web.Request, number=None) -> web.Response:
    """numbers_prime_is_pell_prime_get_0

    Checks whether a given number is a known pell prime number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_prime_is_perfect_get_0(request: web.Request, number=None) -> web.Response:
    """numbers_prime_is_perfect_get_0

    Checks whether a given number is a perfect number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)


async def numbers_prime_is_prime_get_0(request: web.Request, number=None) -> web.Response:
    """numbers_prime_is_prime_get_0

    Checks whether a given number is a known prime number or not.

    :param number: Number to check
    :type number: int

    """
    return web.Response(status=200)
